requirements.txt
fastapi
uvicorn
pandas
PyPDF2
python-docx
python-pptx
openpyxl
